pyproject.toml
[project]
name = "restobox"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12,<3.13"
dependencies = [
    "kornia>=0.8.0",
    "onnx>=1.17.0",
    "onnxconverter-common>=1.14.0",
    "onnxruntime>=1.21.1",
    "onnxscript>=0.2.5",
    "pillow>=11.0.0",
    "pnnx>=20250427",
    "pytorch-msssim>=1.0.0",
    "tdqm>=0.0.1",
    "timm>=1.0.15",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "visdom>=0.2.4",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu128"
